% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitlgl.R
\name{bitlgl}
\alias{bitlgl}
\alias{as_bitlgl}
\alias{as_bitlgl.default}
\alias{as_bitlgl.pkd_bitlgl}
\alias{as_bitlgl.logical}
\title{Bit-packed logical}
\usage{
bitlgl(x = raw(), extra_bits = 0L)

as_bitlgl(x, ...)

\method{as_bitlgl}{default}(x, ...)

\method{as_bitlgl}{pkd_bitlgl}(x, ...)

\method{as_bitlgl}{logical}(x, ...)
}
\arguments{
\item{x}{A \code{\link[=raw]{raw()}} vector of bytes}

\item{extra_bits}{The number of bits from the last byte that
should be considered part of the vector.}

\item{...}{Unused}
}
\value{
A pkd_vctr of class pkd_bitlgl
}
\description{
Bit-packed logical
}
\examples{
bitlgl(0xff)
bitlgl(0x80)
bitlgl(c(0x01, 0xff), extra_bits = 1)

}
