% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgl8.R
\name{bytelgl}
\alias{bytelgl}
\title{Byte logical vector}
\usage{
bytelgl(x)
}
\arguments{
\item{x}{A \code{\link[=raw]{raw()}} vector of bytes with each element representing an
individual logical value. Values of \code{0x00} are considered \code{FALSE};
all other values are considered \code{TRUE}.}
}
\value{
A pkd_vctr of class pkd_lgl8
}
\description{
Byte logical vector
}
