% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbl32.R
\name{dbl32}
\alias{dbl32}
\alias{as_dbl32}
\alias{as_dbl32.default}
\alias{as_dbl32.pkd_dbl32}
\alias{as_dbl32.integer}
\alias{as_dbl32.numeric}
\alias{new_pkd_dbl32}
\title{Signed 32-bit floating-point vector}
\usage{
dbl32(x = raw(), endian = pkd_system_endian())

as_dbl32(x, ...)

\method{as_dbl32}{default}(x, ...)

\method{as_dbl32}{pkd_dbl32}(x, ...)

\method{as_dbl32}{integer}(x, ...)

\method{as_dbl32}{numeric}(x, ...)

new_pkd_dbl32(x)
}
\arguments{
\item{x}{A \code{\link[=raw]{raw()}} vector of bytes with every eight elements representing an
individual floating point value.}

\item{endian}{0 for the most significant byte first (big endian),
1 for the least significant byte first (little endian).}

\item{...}{Unused}
}
\value{
A pkd_vctr of class pkd_dbl32
}
\description{
Signed 32-bit floating-point vector
}
