% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbl64.R
\name{dbl64}
\alias{dbl64}
\alias{as_dbl64}
\alias{as_dbl64.default}
\alias{as_dbl64.pkd_dbl64}
\alias{as_dbl64.integer}
\alias{as_dbl64.numeric}
\alias{new_pkd_dbl64}
\title{Signed 64-bit floating-point vector}
\usage{
dbl64(x = raw(), endian = pkd_system_endian())

as_dbl64(x, ...)

\method{as_dbl64}{default}(x, ...)

\method{as_dbl64}{pkd_dbl64}(x, ...)

\method{as_dbl64}{integer}(x, ...)

\method{as_dbl64}{numeric}(x, ...)

new_pkd_dbl64(x)
}
\arguments{
\item{x}{A \code{\link[=raw]{raw()}} vector of bytes with every eight elements representing an
individual floating point value.}

\item{endian}{0 for the most significant byte first (big endian),
1 for the least significant byte first (little endian).}

\item{...}{Unused}
}
\value{
A pkd_vctr of class pkd_dbl64
}
\description{
Signed 64-bit floating-point vector
}
