% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int16.R
\name{int16}
\alias{int16}
\alias{as_int16}
\alias{as_int16.default}
\alias{as_int16.pkd_int16}
\alias{as_int16.integer}
\alias{as_int16.numeric}
\alias{new_pkd_int16}
\title{Signed 16-bit integer vector}
\usage{
int16(x = raw(), endian = pkd_system_endian())

as_int16(x, ...)

\method{as_int16}{default}(x, ...)

\method{as_int16}{pkd_int16}(x, ...)

\method{as_int16}{integer}(x, ...)

\method{as_int16}{numeric}(x, ...)

new_pkd_int16(x)
}
\arguments{
\item{x}{A \code{\link[=raw]{raw()}} vector of bytes with every two elements representing an
individual unsigned integer value.}

\item{endian}{0 for the most significant byte first (big endian),
1 for the least significant byte first (little endian).}

\item{...}{Unused}
}
\value{
A pkd_vctr of class pkd_int16
}
\description{
Signed 16-bit integer vector
}
