% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int32.R
\name{int32}
\alias{int32}
\alias{as_int32}
\alias{as_int32.default}
\alias{as_int32.pkd_int32}
\alias{as_int32.integer}
\alias{as_int32.numeric}
\alias{new_pkd_int32}
\title{Signed 32-bit integer vector}
\usage{
int32(x = raw(), endian = pkd_system_endian())

as_int32(x, ...)

\method{as_int32}{default}(x, ...)

\method{as_int32}{pkd_int32}(x, ...)

\method{as_int32}{integer}(x, ...)

\method{as_int32}{numeric}(x, ...)

new_pkd_int32(x)
}
\arguments{
\item{x}{A \code{\link[=raw]{raw()}} vector of bytes with every four elements representing an
individual signed integer value.}

\item{endian}{0 for the most significant byte first (big endian),
1 for the least significant byte first (little endian).}

\item{...}{Unused}
}
\value{
A pkd_vctr of class pkd_int32
}
\description{
Signed 32-bit integer vector
}
