% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int64.R
\name{int64}
\alias{int64}
\alias{as_int64}
\alias{as_int64.default}
\alias{as_int64.pkd_int64}
\alias{as_int64.integer}
\alias{as_int64.numeric}
\alias{new_pkd_int64}
\title{Signed 64-bit integer vector}
\usage{
int64(x = raw(), endian = pkd_system_endian())

as_int64(x, ...)

\method{as_int64}{default}(x, ...)

\method{as_int64}{pkd_int64}(x, ...)

\method{as_int64}{integer}(x, ...)

\method{as_int64}{numeric}(x, ...)

new_pkd_int64(x)
}
\arguments{
\item{x}{A \code{\link[=raw]{raw()}} vector of bytes with every eight elements representing an
individual signed integer value.}

\item{endian}{0 for the most significant byte first (big endian),
1 for the least significant byte first (little endian).}

\item{...}{Unused}
}
\value{
A pkd_vctr of class pkd_int64
}
\description{
Signed 64-bit integer vector
}
