% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkd-package.R
\docType{package}
\name{pkd-package}
\alias{pkd}
\alias{pkd-package}
\title{pkd: Compact Integer and Float Vectors}
\description{
Compact representations of integer and float vectors.
}
\author{
\strong{Maintainer}: Dewey Dunnington \email{dewey@fishandwhistle.net} (\href{https://orcid.org/0000-0002-9415-4582}{ORCID})

}
\keyword{internal}
