% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vctr.R
\name{pkd_as_atomic}
\alias{pkd_as_atomic}
\alias{pkd_swap_endian}
\alias{pkd_ensure_endian}
\alias{pkd_system_endian}
\title{Packed vectors}
\usage{
pkd_as_atomic(x)

pkd_swap_endian(x)

pkd_ensure_endian(x, endian = pkd_system_endian())

pkd_system_endian()
}
\arguments{
\item{x}{A pkd_vctr}

\item{endian}{0 for the most significant byte first (XDR),
1 for the least significant byte first.}
}
\description{
Packed vectors
}
