% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vctr.R
\name{pkd_as_r_vector}
\alias{pkd_as_r_vector}
\alias{pkd_swap_endian}
\alias{pkd_ensure_endian}
\alias{pkd_system_endian}
\title{Packed vectors}
\usage{
pkd_as_r_vector(x)

pkd_swap_endian(x)

pkd_ensure_endian(x, endian = pkd_system_endian())

pkd_system_endian()
}
\arguments{
\item{x}{A pkd_vctr}

\item{endian}{0 for the most significant byte first (big endian),
1 for the least significant byte first (little endian).}
}
\description{
Packed vectors
}
