% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uint32.R
\name{uint32}
\alias{uint32}
\alias{as_uint32}
\alias{as_uint32.default}
\alias{as_uint32.pkd_uint32}
\alias{as_uint32.integer}
\alias{as_uint32.numeric}
\alias{new_pkd_uint32}
\title{Unsigned 32-bit integer vector}
\usage{
uint32(x = raw(), endian = pkd_system_endian())

as_uint32(x, ...)

\method{as_uint32}{default}(x, ...)

\method{as_uint32}{pkd_uint32}(x, ...)

\method{as_uint32}{integer}(x, ...)

\method{as_uint32}{numeric}(x, ...)

new_pkd_uint32(x)
}
\arguments{
\item{x}{A \code{\link[=raw]{raw()}} vector of bytes with every four elements representing an
individual unsigned integer value.}

\item{endian}{0 for the most significant byte first (big endian),
1 for the least significant byte first (little endian).}

\item{...}{Unused}
}
\value{
A pkd_vctr of class pkd_uint32
}
\description{
Unsigned 32-bit integer vector
}
