% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uint64.R
\name{uint64}
\alias{uint64}
\alias{as_uint64}
\alias{as_uint64.default}
\alias{as_uint64.pkd_uint64}
\alias{as_uint64.integer}
\alias{as_uint64.numeric}
\alias{new_pkd_uint64}
\title{Unsigned 64-bit integer vector}
\usage{
uint64(x = raw(), endian = pkd_system_endian())

as_uint64(x, ...)

\method{as_uint64}{default}(x, ...)

\method{as_uint64}{pkd_uint64}(x, ...)

\method{as_uint64}{integer}(x, ...)

\method{as_uint64}{numeric}(x, ...)

new_pkd_uint64(x)
}
\arguments{
\item{x}{A \code{\link[=raw]{raw()}} vector of bytes with every eight elements representing an
individual unsigned integer value.}

\item{endian}{0 for the most significant byte first (big endian),
1 for the least significant byte first (little endian).}

\item{...}{Unused}
}
\value{
A pkd_vctr of class pkd_uint64
}
\description{
Unsigned 64-bit integer vector
}
