% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uint8.R
\name{uint8}
\alias{uint8}
\alias{as_uint8}
\alias{as_uint8.default}
\alias{as_uint8.pkd_uint8}
\alias{as_uint8.integer}
\alias{as_uint8.numeric}
\alias{new_pkd_uint8}
\title{Unsigned 8-bit integer vector}
\usage{
uint8(x = raw())

as_uint8(x, ...)

\method{as_uint8}{default}(x, ...)

\method{as_uint8}{pkd_uint8}(x, ...)

\method{as_uint8}{integer}(x, ...)

\method{as_uint8}{numeric}(x, ...)

new_pkd_uint8(x)
}
\arguments{
\item{x}{A \code{\link[=raw]{raw()}} vector of bytes with each element representing an
individual unsigned integer value.}

\item{...}{Unused}
}
\value{
A pkd_vctr of class pkd_uint8
}
\description{
Unsigned 8-bit integer vector
}
